Rename module folders to non plural version
