+++ pd_utils/cli/user_report_cli.py
@@ -3,7 +3,7 @@ Command line handler for UserReport
 """
 from __future__ import annotations
 
-from pd_utils.reports import UserReport
+from pd_utils.report import UserReport
 from pd_utils.util import DateTool
 from pd_utils.util import IOUtil
 from pd_utils.util import RuntimeInit
